+++ src/main/java/homework_5/exp_w_rec/Main.java
@@ -1,0 +1,7 @@
+package homework_5.exp_w_rec;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println(new ExpWRec().run(5.4, 5));
+    }
+}
