--- src/main/java/homework_5/exponentiation_via_recursion/Main.java
@@ -1,7 +1,0 @@
-package homework_5.exponentiation_via_recursion;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println(new ExponentiationViaRecursion().run(5.4, 5));
-    }
-}
